requirements.txt
# Core ML/NLP Libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

# Topic Modeling
gensim>=4.2.0
pyLDAvis>=3.3.0

# Text Processing
spacy>=3.4.0
nltk>=3.7

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
wordcloud>=1.8.0

# MLflow for Model Tracking
mlflow>=2.0.0

# Configuration Management
pyyaml>=6.0
python-dotenv>=0.19.0

# Jupyter for Analysis
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
tqdm>=4.64.0
click>=8.0.0

# Optional: For faster processing
swifter>=1.3.0

# Development
pytest>=7.0.0
black>=22.0.0
